Abstract the representation of a complex of molecules.
Complex of molecules was concrete and not inheritable, where the problem
was in using a graph to index the complex in the `build` function of
`MoleculeComplexBuilder`. The usage of a graph has been moved to its own
protected function so that subclasses of `MoleculeComplexBuilder` can
use the grouping as well.