Remove factory parameters from CML parsing, add a dedicated function for getting a base unit in a unit system, and add object transforms.
The extension function for parsing a CML string and importing its
constructs no longer accepts factory functions for atoms and bonds,
since factory functions do not really solve the problem of constructing
instances of subclasses in this case. Instead, a general interface for
transforms is added to allow an implementing class to define the
conversion of a deserialized molecule complex to a molecule complex of a
different type.